fix: conn. test was unreliable && made server fall
- Memleaks in the test are now checked with -fanitize instead of valgrind,
this fixed important issues.
- Server doesn't allow select to fail due to excess connection attempts
anymore. It just doesn't connect and keeps running, this is safer.
- Signal now handles all interruption signals instead of only SIGINT

modified:   Makefile
modified:   integration_tests/Makefile
modified:   srcs/Server.cpp
modified:   srcs/Server.hpp
new file:   integration_tests/connections.py
deleted:    integration_tests/1024_connections.py
