+++ Makefile
@@ -1,6 +1,8 @@
 PORT ?= 6667
 NAME = ircserv
+NAME2 = ircserv
 FLAGS = -Wall -Wextra -Werror -std=c++98 -g3 -MMD -MP
+MEMFLAGS = -Wall -Wextra -Werror -std=c++98 -g3 -MMD -MP -fsanitize=address
 CXX = c++
 PATH_OBJS = ./objects/
 SRC_DIR = srcs/
@@ -21,6 +23,9 @@ all: $(NAME)
 run: all
 	valgrind --leak-check=full --show-leak-kinds=all --track-origins=yes ./$(NAME) $(PORT) $(PASS)
 
+quickmemtest: fclean $(NAME2)
+	./$(NAME) $(PORT) $(PASS)
+
 test_connections:
 	@make test1 PORT=$(PORT) -C integration_tests/
 
@@ -37,6 +42,18 @@ $(MAIN_OBJ): $(SRC_DIR)$(MAIN).cpp
 	@mkdir -p $(PATH_OBJS)
 	$(CXX) $(FLAGS) -c $< -o $@
 
+$(NAME2): $(OBJS) $(MAIN_OBJ)
+	@mkdir -p $(dir $@)
+	$(CXX) $(MEMFLAGS) $^ -o $@
+	@touch $@
+
+$(PATH_OBJS)%.o: $(SRC_DIR)%.cpp
+	@mkdir -p $(PATH_OBJS)
+	$(CXX) $(MEMFLAGS) -c $< -o $@
+
+$(MAIN_OBJ): $(SRC_DIR)$(MAIN).cpp
+	@mkdir -p $(PATH_OBJS)
+	$(CXX) $(MEMFLAGS) -c $< -o $@
 clean:
 	rm -rf $(PATH_OBJS)
